+++ packages/mixte/src/pick/index.ts
@@ -11,19 +11,19 @@ import { toArray } from 'mixte';
  * pick({ a: 1, b: 2, c: 3 }, ['a', 'c']); // -> { a: 1, c: 3 }
  * pick({ a: 1, b: 2, c: 3 }, 'a'); // -> { a: 1 }
  */
-export function pick<T extends object, K extends keyof T>(obj: T, keys: K[] | K): Pick<T, K> {
+export function pick<T extends object, K extends keyof T>(obj: T, keys: K[] | readonly K[] | K): Pick<T, K> {
   if (!obj) return {} as Pick<T, K>;
 
-  const result = {} as Pick<T, K>;
+  const result: Record<string, any> = {};
   const keysArray = toArray(keys);
 
   for (const key of keysArray) {
-    if (key in obj) {
-      result[key] = obj[key];
+    if (key as keyof T in obj) {
+      result[key as string] = obj[key as keyof T];
     }
   }
 
-  return result;
+  return result as Pick<T, K>;
 }
 
 /**
@@ -37,14 +37,14 @@ export function pick<T extends object, K extends keyof T>(obj: T, keys: K[] | K)
  * omit({ a: 1, b: 2, c: 3 }, ['a', 'c']); // -> { b: 2 }
  * omit({ a: 1, b: 2, c: 3 }, 'a'); // -> { b: 2, c: 3 }
  */
-export function omit<T extends object, K extends keyof T>(obj: T, keys: K[] | K): Omit<T, K> {
+export function omit<T extends object, K extends keyof T>(obj: T, keys: K[] | readonly K[] | K): Omit<T, K> {
   if (!obj) return {} as Omit<T, K>;
 
   const result = { ...obj };
   const keysArray = toArray(keys);
 
   for (const key of keysArray) {
-    delete result[key];
+    delete result[key as keyof T];
   }
 
   return result as Omit<T, K>;