+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-  - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `Welcome` 组件
+  - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `Welcome` 组件 *
+  - 💄 [mixte] 优化 `pick` 和 `omit` 函数类型 *
 
 ## [v3.4.0-beta.2]
   - 📅 2025-03-16